dogs_breed_classification/requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
Bottleneck==1.2.1
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
Markdown==2.6.11
numpy==1.14.2
pandas==0.22.0
Pillow==5.4.1
protobuf==3.5.2.post1
python-dateutil==2.7.2
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
Werkzeug==0.14.1

